Enum glean_core::metrics::TimeUnit

#[repr(i32)]
pub enum TimeUnit {
    Nanosecond = 0,
    Microsecond = 1,
    Millisecond = 2,
    Second = 3,
    Minute = 4,
    Hour = 5,
    Day = 6,
}

Different resolutions supported by the time related metric types (e.g. DatetimeMetric).

Variants

Nanosecond = 0

Truncate to nanosecond precision.

Microsecond = 1

Truncate to microsecond precision.

Millisecond = 2

Truncate to millisecond precision.

Second = 3

Truncate to second precision.

Minute = 4

Truncate to minute precision.

Hour = 5

Truncate to hour precision.

Day = 6

Truncate to day precision.

Implementations

impl TimeUnit

pub fn format_pattern(self) -> &'static str

Formats the given time unit, truncating the time if needed.

pub fn duration_convert(self, duration: Duration) -> u64

Converts a duration to the requested time unit.

Arguments

duration - the duration to convert.

Returns

The integer representation of the converted duration.

pub fn as_nanos(self, duration: u64) -> u64

Converts a duration in the given unit to nanoseconds.

Arguments

duration - the duration to convert.

Returns

The integer representation of the nanosecond duration.