Struct glean_core::metrics::EventMetric

pub struct EventMetric { /* private fields */ }

An event metric.

Events allow recording of e.g. individual occurences of user actions, say every time a view was open and from where. Each time you record an event, it records a timestamp, the event’s name and a set of custom values.

Implementations

impl EventMetric

pub fn new(meta: CommonMetricData, allowed_extra_keys: Vec<String>) -> Self

Creates a new event metric.

pub fn record(&self, extra: HashMap<String, String>)

Records an event.

Arguments

extra - A HashMap of (key, value) pairs. Keys must be one of the allowed extra keys. If any key is not allowed, an error is reported and no event is recorded.

pub fn record_with_time(&self, timestamp: u64, extra: HashMap<String, String>)

Record a new event with a provided timestamp.

It’s the caller’s responsibility to ensure the timestamp comes from the same clock source.

Arguments

timestamp - The event timestamp, in milliseconds.

extra - A HashMap of (key, value) pairs. Keys must be one of the allowed extra keys. If any key is not allowed, an error is reported and no event is recorded.

pub fn test_get_value( &self, ping_name: Option<String> ) -> Option<Vec<RecordedEvent>>

Test-only API (exported for FFI purposes).

Get the vector of currently stored events for this event metric.

This doesn’t clear the stored value.

Arguments

ping_name - the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

pub fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

Returns

The number of errors reported.