Struct glean_core::metrics::MemoryDistributionMetric

pub struct MemoryDistributionMetric { /* fields omitted */ }

A memory distribution metric.

Memory distributions are used to accumulate and store memory sizes.

Implementations

impl MemoryDistributionMetric

pub fn new(meta: CommonMetricData, memory_unit: MemoryUnit) -> Self

Create a new memory distribution metric.

pub fn accumulate(&self, glean: &Glean, sample: u64)

Accumulates the provided sample in the metric.

Arguments

sample - The sample to be recorded by the metric. The sample is assumed to be in the configured memory unit of the metric.

Notes

Values bigger than 1 Terabyte (240 bytes) are truncated and an ErrorType::InvalidValue error is recorded.

pub fn accumulate_samples_signed(&self, glean: &Glean, samples: Vec<i64>)

Accumulates the provided signed samples in the metric.

This is required so that the platform-specific code can provide us with 64 bit signed integers if no u64 comparable type is available. This will take care of filtering and reporting errors for any provided negative sample.

Please note that this assumes that the provided samples are already in the "unit" declared by the instance of the implementing metric type (e.g. if the implementing class is a [MemoryDistributionMetricType] and the instance this method was called on is using [MemoryUnit.Kilobyte], then samples are assumed to be in that unit).

Arguments

samples - The vector holding the samples to be recorded by the metric.

Notes

Discards any negative value in samples and report an ErrorType::InvalidValue for each of them. Values bigger than 1 Terabyte (240 bytes) are truncated and an ErrorType::InvalidValue error is recorded.

pub fn test_get_value(
&self,
glean: &Glean,
storage_name: &str
) -> Option<DistributionData>

Test-only API (exported for FFI purposes).

Get the currently stored value as an integer.

This doesn't clear the stored value.

pub fn test_get_value_as_json_string(
&self,
glean: &Glean,
storage_name: &str
) -> Option<String>

Test-only API (exported for FFI purposes).

Get the currently-stored histogram as a JSON String of the serialized value.

This doesn't clear the stored value.