Struct glean_core::metrics::TimingDistributionMetric

pub struct TimingDistributionMetric { /* fields omitted */ }

A timing distribution metric.

Timing distributions are used to accumulate and store time measurement, for analyzing distributions of the timing data.

Methods

impl TimingDistributionMetric

pub fn new(meta: CommonMetricData, time_unit: TimeUnit) -> Self

Create a new timing distribution metric.

pub fn set_start(&mut self, glean: &Glean, start_time: u64) -> TimerId

Start tracking time for the provided metric.

This records an error if it’s already tracking time (i.e. start was already called with no corresponding [stop]): in that case the original start time will be preserved.

Arguments

start_time - Timestamp in nanoseconds.

Return value

Returns a unique TimerId for the new timer.

pub fn set_stop_and_accumulate(
&mut self,
glean: &Glean,
id: TimerId,
stop_time: u64
)

Stop tracking time for the provided metric and associated timer id. Add a count to the corresponding bucket in the timing distribution. This will record an error if no start was called.

Arguments

id - The TimerId to associate with this timing. This allows for concurrent timing of events associated with different ids to the same timespan metric.

stop_time - Timestamp in nanoseconds.

pub fn cancel(&mut self, id: TimerId)

Abort a previous set_start call. No error is recorded if no set_start was called.

Arguments

id - The TimerId to associate with this timing. This allows for concurrent timing of events associated with different ids to the same timing distribution metric.

pub fn accumulate_samples_signed(&mut self, glean: &Glean, samples: Vec<i64>)

Accumulates the provided signed samples in the metric.

This is required so that the platform-specific code can provide us with 64 bit signed integers if no u64 comparable type is available. This will take care of filtering and reporting errors for any provided negative sample.

Please note that this assumes that the provided samples are already in the "unit" declared by the instance of the implementing metric type (e.g. if the implementing class is a [TimingDistributionMetricType] and the instance this method was called on is using [TimeUnit.Second], then samples are assumed to be in that unit).

Arguments

samples - The vector holding the samples to be recorded by the metric.

Notes

Discards any negative value in samples and report an ErrorType::InvalidValue for each of them.

pub fn test_get_value(
&self,
glean: &Glean,
storage_name: &str
) -> Option<Histogram>

Test-only API (exported for FFI purposes).

Get the currently stored value as an integer.

This doesn't clear the stored value.

pub fn test_get_value_as_json_string(
&self,
glean: &Glean,
storage_name: &str
) -> Option<String>

Test-only API (exported for FFI purposes).

Get the currently-stored histogram as a JSON String of the serialized value.

This doesn't clear the stored value.