Struct glean_core::storage::StorageManager

pub struct StorageManager;

Snapshot metrics from the underlying database.

Implementations

impl StorageManager

pub fn snapshot(
&self,
storage: &Database,
store_name: &str,
clear_store: bool
) -> Option<String>

Snapshots the given store and optionally clear it.

Arguments

storage - the database to read from.

store_name - the store to snapshot.

clear_store - whether to clear the data after snapshotting.

Returns

The stored data in a string encoded as JSON. If no data for the store exists, None is returned.

pub fn snapshot_as_json(
&self,
storage: &Database,
store_name: &str,
clear_store: bool
) -> Option<JsonValue>

Snapshots the given store and optionally clear it.

Arguments

storage - the database to read from.

store_name - the store to snapshot.

clear_store - whether to clear the data after snapshotting.

Returns

A JSON representation of the stored data. If no data for the store exists, None is returned.

pub fn snapshot_metric(
&self,
storage: &Database,
store_name: &str,
metric_id: &str,
metric_lifetime: Lifetime
) -> Option<Metric>

Gets the current value of a single metric identified by name.

Arguments

storage - The database to get data from.

store_name - The store name to look into.

metric_id - The full metric identifier.

Returns

The decoded metric or None if no data is found.

pub fn snapshot_metric_for_test(
&self,
storage: &Database,
store_name: &str,
metric_id: &str,
metric_lifetime: Lifetime
) -> Option<Metric>

Gets the current value of a single metric identified by name.

Use this API, rather than snapshot_metric within the testing API, so that the usage will be reported in coverage, if enabled.

Arguments

storage - The database to get data from.

store_name - The store name to look into.

metric_id - The full metric identifier.

Returns

The decoded metric or None if no data is found.

pub fn snapshot_experiments_as_json(
&self,
storage: &Database,
store_name: &str
) -> Option<JsonValue>

Snapshots the experiments.

Arguments

storage - The database to get data from.

store_name - The store name to look into.

Returns

A JSON representation of the experiment data, in the following format:

{
 "experiment-id": {
   "branch": "branch-id",
   "extra": {
     "additional": "property",
     // ...
   }
 }
}

If no data for the store exists, None is returned.