Struct glean_core::ClientInfoMetrics

pub struct ClientInfoMetrics {
    pub app_build: String,
    pub app_display_version: String,
    pub app_build_date: Datetime,
    pub architecture: String,
    pub os_version: String,
    pub channel: Option<String>,
    pub android_sdk_version: Option<String>,
    pub windows_build_number: Option<i64>,
    pub device_manufacturer: Option<String>,
    pub device_model: Option<String>,
    pub locale: Option<String>,
}

Metrics included in every ping as client_info.

Fields
app_build: String

The build identifier generated by the CI system (e.g. “1234/A”).

app_display_version: String

The user visible version string (e.g. “1.0.3”).

app_build_date: Datetime

The app’s build date

architecture: String

The architecture of the device (e.g. “arm”, “x86”).

os_version: String

The name of the operating system (e.g. “Linux”, “Android”, “iOS”).

channel: Option<String>

The product-provided release channel (e.g. “beta”).

android_sdk_version: Option<String>

The Android specific SDK version of the software running on this hardware device (e.g. “23”).

windows_build_number: Option<i64>

The Windows specific OS build version (e.g. 19043)

device_manufacturer: Option<String>

The manufacturer of the device the application is running on. Not set if the device manufacturer can’t be determined (e.g. on Desktop).

device_model: Option<String>

The model of the device the application is running on. On Android, this is Build.MODEL, the user-visible marketing name, like “Pixel 2 XL”. Not set if the device model can’t be determined (e.g. on Desktop).

locale: Option<String>

The locale of the application during initialization (e.g. “es-ES”). If the locale can’t be determined on the system, the value is “und”, to indicate “undetermined”.

Implementations

impl ClientInfoMetrics

Metrics included in every ping as client_info.

pub fn unknown() -> Self

Creates the client info with dummy values for all.