Trait glean_core::traits::Counter

pub trait Counter {
    // Required methods
    fn add(&self, amount: i32);
    fn test_get_value<'a, S: Into<Option<&'a str>>>(
        &self,
        ping_name: S
    ) -> Option<i32>;
    fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32;
}

A description for the CounterMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required Methods

fn add(&self, amount: i32)

Increases the counter by amount.

Arguments

amount - The amount to increase by. Should be positive.

Notes

Logs an error if the amount is 0 or negative.

fn test_get_value<'a, S: Into<Option<&'a str>>>( &self, ping_name: S ) -> Option<i32>

Exported for test purposes.

Gets the currently stored value as an integer.

This doesn’t clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

Returns

The number of errors reported.

Object Safety

This trait is not object safe.