Trait glean_core::traits::CustomDistribution

pub trait CustomDistribution {
    fn accumulate_samples_signed(&self, samples: Vec<i64>);

    fn test_get_value(&self, storage_name: &str);

    fn test_get_value_as_json_string(
        &self, 
        storage_name: &str
    ) -> Option<String>;
}

A description for the CustomDistributionMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn accumulate_samples_signed(&self, samples: Vec<i64>)

Accumulates the provided signed samples in the metric.

This is required so that the platform-specific code can provide us with 64 bit signed integers if no u64 comparable type is available. This will take care of filtering and reporting errors for any provided negative sample.

Arguments

samples - The vector holding the samples to be recorded by the metric.

Notes

Discards any negative value in samples and report an ErrorType::InvalidValue for each of them.

fn test_get_value(&self, storage_name: &str)

Exported for test purposes.

Gets the currently stored histogram.

This doesn't clear the stored value.

fn test_get_value_as_json_string(&self, storage_name: &str) -> Option<String>

Exported for test purposes.

Gets the currently stored histogram as a JSON String of the serialized value.

This doesn't clear the stored value.

Loading content...