Trait glean_core::traits::Datetime

pub trait Datetime: TestGetValue<Datetime> {
    // Required methods
    fn set(&self, value: Option<Datetime>);
    fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32;
}

A description for the DatetimeMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required Methods

fn set(&self, value: Option<Datetime>)

Sets the metric to a date/time which including the timezone offset.

Arguments

value - Some Datetime, with offset, to set the metric to. If None, the current local time is used.

fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

Returns

The number of errors reported.