Trait glean_core::traits::Jwe

pub trait Jwe {
    fn set_with_compact_representation<S: Into<String>>(&self, value: S);

    fn set<S: Into<String>>(
        &self, 
        header: S, 
        key: S, 
        init_vector: S, 
        cipher_text: S, 
        auth_tag: S
    );

    fn test_get_value(&self, storage_name: &str) -> Option<String>;

    fn test_get_value_as_json_string(
        &self, 
        storage_name: &str
    ) -> Option<String>;
}

A description for the JweMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn set_with_compact_representation<S: Into<String>>(&self, value: S)

Sets to the specified JWE value.

Arguments

value - the compact representation of a JWE value.

fn set<S: Into<String>>(
&self,
header: S,
key: S,
init_vector: S,
cipher_text: S,
auth_tag: S
)

Builds a JWE value from its elements and set to it.

Arguments

header - the JWE Protected Header element.

key - the JWE Encrypted Key element.

init_vector - the JWE Initialization Vector element.

cipher_text - the JWE Ciphertext element.

auth_tag - the JWE Authentication Tag element.

fn test_get_value(&self, storage_name: &str) -> Option<String>

Exported for test purposes.

Gets the currently stored value as a string.

This doesn't clear the stored value.

fn test_get_value_as_json_string(&self, storage_name: &str) -> Option<String>

Exported for test purposes.

Gets the currently stored JWE as a JSON String of the serialized value.

This doesn't clear the stored value.

Loading content...