Trait glean_core::traits::String

pub trait String {
    fn set<S: Into<String>>(&self, value: S);

    fn test_get_value<'a, S: Into<Option<&'a str>>>(
        &self, 
        ping_name: S
    ) -> Option<String>;

    fn test_get_num_recorded_errors<'a, S: Into<Option<&'a str>>>(
        &self, 
        error: ErrorType, 
        ping_name: S
    ) -> i32;
}

A description for the StringMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn set<S: Into<String>>(&self, value: S)

Sets to the specified value.

Arguments

value - The string to set the metric to.

Notes

Truncates the value if it is longer than MAX_LENGTH_VALUE bytes and logs an error.

fn test_get_value<'a, S: Into<Option<&'a str>>>(
&self,
ping_name: S
) -> Option<String>

Exported for test purposes.

Gets the currently stored value as a string.

This doesn’t clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_num_recorded_errors<'a, S: Into<Option<&'a str>>>(
&self,
error: ErrorType,
ping_name: S
) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The number of errors reported.