Module glean_core::upload::ffi_upload_result

Result values of attempted ping uploads encoded for FFI use.

In a perfect world this would live in glean-ffi, but because we also want to convert from pure integer values to a proper Rust enum using Rust's From and Into trait, we need to have it in this crate (The coherence rules don't allow to implement an external trait for an external type).

Due to restrictions of cbindgen they are re-defined in glean-core/ffi/src/upload.rs.

NOTE: THEY MUST BE THE SAME ACROSS BOTH FILES!

Constants

UPLOAD_RESULT_HTTP_STATUS
A HTTP response code.

UPLOAD_RESULT_RECOVERABLE
A recoverable error.

UPLOAD_RESULT_UNRECOVERABLE
An unrecoverable error.