Struct glean_core::upload::PingRequest

pub struct PingRequest {
    pub document_id: String,
    pub path: String,
    pub body: Vec<u8>,
    pub headers: HeaderMap,
    pub body_has_info_sections: bool,
    pub ping_name: String,
    pub uploader_capabilities: Vec<String>,
}

Represents a request to upload a ping.

Fields
document_id: String

The Job ID to identify this request, this is the same as the ping UUID.

path: String

The path for the server to upload the ping to.

body: Vec<u8>

The body of the request, as a byte array. If gzip encoded, then the headers list will contain a Content-Encoding header with the value gzip.

headers: HeaderMap

A map with all the headers to be sent with the request.

body_has_info_sections: bool

Whether the body has {client|ping}_info sections.

ping_name: String

The ping’s name. Likely also somewhere in path.

uploader_capabilities: Vec<String>

The capabilities required during this ping’s upload.

Implementations

impl PingRequest

pub fn builder(language_binding_name: &str, body_max_size: usize) -> Builder

Creates a new builder-style structure to help build a PingRequest.

Arguments

language_binding_name - The name of the language used by the binding that instantiated this Glean instance. This is used to build the X-Telemetry-Agent header value.

body_max_size - The maximum size in bytes the compressed ping body may have to be eligible for upload.

pub fn is_deletion_request(&self) -> bool

Verifies if current request is for a deletion-request ping.

pub fn pretty_body(&self) -> Option<String>

Decompresses and pretty-format the ping payload

Should be used for logging when required. This decompresses the payload in memory.