Crate glean_core

Glean is a modern approach for recording and sending Telemetry data.

It’s in use at Mozilla.

All documentation can be found online:

The Glean SDK Book

Modules

metrics
The different metric types supported by the Glean SDK to handle data.

ping
Ping collection, assembly & submission.

storage
Storage snapshotting.

traits
API definitions for the different metric types supported by the Glean SDK.

upload
Manages the pending pings queue and directory.

Structs

CommonMetricData
The common set of data shared across all different metric types.

Configuration
The Glean configuration.

Error
A specialized Error type for this crate’s operations.

Glean
The object holding meta information about a Glean instance.

Enums

ErrorKind
A list enumerating the categories of errors in this crate.

ErrorType
The possible error types for metric recording. Note: the cases in this enum must be kept in sync with the ones in the platform-specific code (e.g. ErrorType.kt) and with the metrics in the registry files.

HistogramType
Different kinds of histograms.

Lifetime
The supported metrics’ lifetimes.

Functions

get_timestamp_ms
Returns a timestamp corresponding to “now” with millisecond precision.

global_glean
Gets a reference to the global Glean object.

setup_glean
Sets or replaces the global Glean object.

test_get_num_recorded_errors
Gets the number of recorded errors for the given metric and error type.

Type Definitions

Result
A specialized Result type for this crate’s operations.