Crate glean_core

Glean is a modern approach for recording and sending Telemetry data.

It's in use at Mozilla.

All documentation can be found online:

The Glean SDK Book

Modules

metrics
The different metric types supported by the Glean SDK to handle data.

ping
Ping collection, assembly & submission.

storage
Storage snapshotting.

Structs

CommonMetricData
The common set of data shared across all different metric types.

Configuration
The Glean configuration.

Error
A specialized Error type for this crate's operations.

Glean
The object holding meta information about a Glean instance.

Enums

ErrorType
The possible error types for metric recording. Note: the cases in this enum must be kept in sync with the ones in the platform-specific code (e.g. ErrorType.kt) and with the metrics in the registry files.

Lifetime
The supported metrics' lifetimes.

Functions

global_glean
Get a reference to the global Glean object.

setup_glean
Set or replace the global Glean object.

test_get_num_recorded_errors
Get the number of recorded errors for the given metric and error type.

Type Definitions

Result
A specialized Result type for this crate's operations.