Module glean_core::metrics

The different metric types supported by the Glean SDK to handle data.

Structs

BooleanMetric

A boolean metric.

CounterMetric

A counter metric.

CustomDistributionMetric

A custom distribution metric.

Datetime

Representation of a date, time and timezone.

DatetimeMetric

A datetime metric.

DenominatorMetric

A Denominator metric (a kind of count shared among Rate metrics).

DistributionData

A snapshot of all buckets and the accumulated sum of a distribution.

EventMetric

An event metric.

LabeledMetric

A labeled metric.

MemoryDistributionMetric

A memory distribution metric.

MetricsDisabledConfig

Represents a list of metrics and their associated disabled property from the remote-settings configuration store. The expected format of this data is stringified JSON in the following format:

NumeratorMetric

Developer-facing API for recording rate metrics with external denominators.

PingType

Stores information about a ping.

QuantityMetric

A quantity metric.

Rate

A rate value as given by its numerator and denominator.

RateMetric

A rate metric.

RecordedEvent

Represents the recorded data for a single event.

RecordedExperiment

Deserialized experiment data.

StringListMetric

A string list metric.

StringMetric

A string metric.

TextMetric

A text metric.

TimerId

Identifier for a running timer.

TimespanMetric

A timespan metric.

TimingDistributionMetric

A timing distribution metric.

UrlMetric

A URL metric.

UuidMetric

An UUID metric.

Enums

HistogramType

Different kinds of histograms.

MemoryUnit

Different resolutions supported by the memory related metric types (e.g. MemoryDistributionMetric).

Metric

The available metrics.

TimeUnit

Different resolutions supported by the time related metric types (e.g. DatetimeMetric).

Traits

MetricType

A MetricType describes common behavior across all metrics.

Type Definitions

LabeledBoolean

A labeled boolean.

LabeledCounter

A labeled counter.

LabeledString

A labeled string.