Module glean_core::metrics

The different metric types supported by the Glean SDK to handle data.

Structs

BooleanMetric

A boolean metric.

CounterMetric

A counter metric.

CustomDistributionMetric

A custom distribution metric.

DatetimeMetric

A datetime metric.

DenominatorMetric

A Denominator metric (a kind of count shared among Rate metrics).

DistributionData

A snapshot of all buckets and the accumulated sum of a distribution.

EventMetric

An event metric.

JweMetric

A JWE metric.

LabeledMetric

A labeled metric.

MemoryDistributionMetric

A memory distribution metric.

PingType

Stores information about a ping.

QuantityMetric

A quantity metric.

RateMetric

A rate metric.

RecordedEvent

Represents the recorded data for a single event.

StringListMetric

A string list metric.

StringMetric

A string metric.

TimespanMetric

A timespan metric.

TimingDistributionMetric

A timing distribution metric.

UrlMetric

A URL metric.

UuidMetric

An UUID metric.

Enums

HistogramType

Different kinds of histograms.

MemoryUnit

Different resolutions supported by the memory related metric types (e.g. MemoryDistributionMetric).

Metric

The available metrics.

TimeUnit

Different resolutions supported by the time related metric types (e.g. DatetimeMetric).

Traits

MetricType

A MetricType describes common behavior across all metrics.

Type Definitions

Datetime

A datetime type.

TimerId

Identifier for a running timer.