Module glean_core::metrics

The different metric types supported by the Glean SDK to handle data.

Structs

BooleanMetric
A boolean metric.

CounterMetric
A counter metric.

CustomDistributionMetric
A custom distribution metric.

DatetimeMetric
A datetime metric.

EventMetric
An event metric.

LabeledMetric
A labeled metric.

MemoryDistributionMetric
A memory distribution metric.

PingType
Stores information about a ping.

QuantityMetric
A quantity metric.

StringListMetric
A string list metric.

StringMetric
A string metric.

TimespanMetric
A timespan metric.

TimingDistributionMetric
A timing distribution metric.

UuidMetric
An UUID metric.

Enums

HistogramType
Different kinds of histograms.

MemoryUnit
Different resolutions supported by the memory related metric types (e.g. MemoryDistributionMetric).

Metric
The available metrics.

TimeUnit
Different resolutions supported by the time related metric types (e.g. DatetimeMetric).

Traits

MetricType
A MetricType describes common behavior across all metrics.

Functions

combine_base_identifier_and_label
Combines a metric's base identifier and label

dynamic_label
Validate a dynamic label, changing it to OTHER_LABEL if it's invalid.

strip_label
Strips the label off of a complete identifier

Type Definitions

TimerId
Identifier for a running timer.