Trait glean_core::metrics::MetricType

pub trait MetricType {
    fn meta(&self) -> &CommonMetricDataInternal;

    fn with_name(&self, _name: String) -> Self
    where
        Self: Sized,
    { ... }
    fn with_dynamic_label(&self, _label: String) -> Self
    where
        Self: Sized,
    { ... }
    fn should_record(&self, glean: &Glean) -> bool { ... }
}

A MetricType describes common behavior across all metrics.

Required Methods

fn meta(&self) -> &CommonMetricDataInternal

Access the stored metadata

Provided Methods

fn with_name(&self, _name: String) -> Selfwhere
Self: Sized,

Create a new metric from this with a new name.

fn with_dynamic_label(&self, _label: String) -> Selfwhere
Self: Sized,

Create a new metric from this with a specific label.

fn should_record(&self, glean: &Glean) -> bool

Whether this metric should currently be recorded

This depends on the metrics own state, as determined by its metadata, and whether upload is enabled on the Glean object.