Trait glean_core::traits::Datetime

pub trait Datetime {
    fn set_with_details(
        &self, 
        year: i32, 
        month: u32, 
        day: u32, 
        hour: u32, 
        minute: u32, 
        second: u32, 
        nano: u32, 
        offset_seconds: i32
    );

    fn set(&self, value: Option<Datetime>);

    fn test_get_value_as_string(&self, storage_name: &str) -> Option<String>;
}

A description for the DatetimeMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn set_with_details(
&self,
year: i32,
month: u32,
day: u32,
hour: u32,
minute: u32,
second: u32,
nano: u32,
offset_seconds: i32
)

Sets the metric to a date/time including the timezone offset.

Arguments

year - the year to set the metric to.

month - the month to set the metric to (1-12).

day - the day to set the metric to (1-based).

hour - the hour to set the metric to.

minute - the minute to set the metric to.

second - the second to set the metric to.

nano - the nanosecond fraction to the last whole second.

offset_seconds - the timezone difference, in seconds, for the Eastern Hemisphere. Negative seconds mean Western Hemisphere.

fn set(&self, value: Option<Datetime>)

Sets the metric to a date/time which including the timezone offset.

Arguments

value - Some date/time value, with offset, to set the metric to. If none, the current local time is used.

fn test_get_value_as_string(&self, storage_name: &str) -> Option<String>

Exported for test purposes.

Gets the currently stored value as a String.

The precision of this value is truncated to the time_unit precision.

This doesn't clear the stored value.

Loading content...