Module glean_core::upload

Manages the pending pings queue and directory.

Keeps track of pending pings, loading any unsent ping from disk on startup;

Exposes get_upload_task API for the platform layer to request next upload task;

Exposes process_ping_upload_response API to check the HTTP response from the ping upload and either delete the corresponding ping from disk or re-enqueue it for sending.

Structs

PingMetadata

A ping’s metadata, as (optionally) represented on disk.

PingPayload

A representation of the data extracted from a ping file,

PingRequest

Represents a request to upload a ping.

PingUploadManager

Manages the pending pings queue and directory.

Enums

PingUploadTask

An enum representing the possible upload tasks to be performed by an uploader.

UploadResult

The result of an attempted ping upload.

UploadTaskAction

Communication back whether the uploader loop should continue.

Functions

chunked_log_info

Logs payload in one go (all other OS).

Type Aliases

HeaderMap

A representation for request headers.