Module glean_core::upload

Manages the pending pings queue and directory.

Keeps track of pending pings, loading any unsent ping from disk on startup;

Exposes get_upload_task API for the platform layer to request next upload task;

Exposes process_ping_upload_response API to check the HTTP response from the ping upload and either delete the corresponding ping from disk or re-enqueue it for sending.

Modules

ffi_upload_result
Result values of attempted ping uploads encoded for FFI use.

Structs

PingRequest
Represents a request to upload a ping.

PingUploadManager
Manages the pending pings queue and directory.

Enums

PingUploadTask
When asking for the next ping request to upload, the requester may receive one out of three possible tasks.

UploadResult
The result of an attempted ping upload.

Functions

chunked_log_info
Log payload in one go (all other OS).

Type Definitions

HeaderMap
A representation for request headers.